+++ adhocracy4/polls/static/PollDashboard/EditPollOpenQuestion.jsx
@@ -31,6 +31,21 @@ export const EditPollOpenQuestion = React.forwardRef((props, ref) => {
           {hasHelptext
             ? <HelptextForm id={props.id} question={props.question} onHelptextChange={props.onHelptextChange} errors={props.errors} />
             : null}
+
+          <div className="form-check">
+            <label className="form-check__label" htmlFor={'id_questions-' + props.id + '-is_confidential'}>
+              <input
+                type="checkbox"
+                id={'id_questions-' + props.id + '-is_confidential'}
+                name={'questions-' + props.id + '-is_confidential'}
+                checked={props.question.is_confidential || false}
+                onChange={(e) => { props.onConfidentialChange(e.target.checked) }}
+              />
+              &nbsp;
+              {django.gettext('Do not display answers publicly')}
+            </label>
+          </div>
+
           <button
             className={'btn ' + (hasHelptext ? 'editpoll__btn--dark' : 'editpoll__btn--question')}
             onClick={() => setHasHelptext(!hasHelptext)}